add run_docker and restructure

+++ c3-project/helper.cpp
@@ -50,7 +50,7 @@ double getDistance(Point p1, Point p2){
 double minDistance(Point p1, vector<Point> points){
 	if(points.size() > 0){
 		double dist = getDistance(p1, points[0]);
-		for(int index = 1; index < points.size(); index++){
+		for(size_t index = 1U; index < points.size(); index++){
 			double newDist = getDistance(p1, points[index]);
 			if( newDist < dist)
 				dist = newDist;
@@ -90,8 +90,8 @@ void renderRay(pcl::visualization::PCLVisualizer::Ptr& viewer, Point p1, Point p
 
 void renderPath(pcl::visualization::PCLVisualizer::Ptr& viewer, const PointCloudT::Ptr& cloud, std::string name, Color color){
 
-	int previous = 0;
-	for(int index = previous+1; index < cloud->points.size(); index++){
+	size_t previous = 0U;
+	for(size_t index = previous+1U; index < cloud->points.size(); index++){
 		renderRay(viewer, Point(cloud->points[previous].x, cloud->points[previous].y, 0), Point(cloud->points[index].x, cloud->points[index].y, 0), name+to_string(previous), color);
 		previous++;
 	}
@@ -102,11 +102,11 @@ void renderPath(pcl::visualization::PCLVisualizer::Ptr& viewer, const PointCloud
 Eigen::Quaternionf getQuaternion(float theta)
 {
 	Eigen::Matrix3f rotation_mat;
-	rotation_mat << 
+	rotation_mat <<
 	cos(theta), -sin(theta), 0,
 	sin(theta),  cos(theta), 0,
 	0, 			 0, 		 1;
-	
+
 	Eigen::Quaternionf q(rotation_mat);
 	return q;
 }
